requirements/local.txt
# Local development dependencies go here
-r base.txt
coverage
django-discover-runner
django-debug-toolbar
Sphinx
mock
factory_boy
selenium